+++ sdl_android_lib/src/com/smartdevicelink/proxy/SdlProxyBase.java
@@ -778,7 +778,6 @@ public abstract class SdlProxyBase<proxyListenerType extends IProxyListenerBase>
 	{		
 		Intent sendIntent = createBroadcastIntent();
 		Intent sendIntent2 = createBroadcastIntent();
-
 		HttpURLConnection urlConnection = null;
 		boolean bLegacy = false;
 		
@@ -2521,7 +2520,6 @@ public abstract class SdlProxyBase<proxyListenerType extends IProxyListenerBase>
 					// OnSystemRequest
 					
 					final OnSystemRequest msg = new OnSystemRequest(hash);
-					
 					if ( (msg.getUrl() != null) &&
 						 (msg.getRequestType() == RequestType.PROPRIETARY) &&
 						 (msg.getFileType() == FileType.JSON) )
@@ -4281,7 +4279,7 @@ public abstract class SdlProxyBase<proxyListenerType extends IProxyListenerBase>
 	 * @param iLength - The total length of the file being sent.
 	 * @throws SdlException
 	*/
-	public void PutFileStream(InputStream is, String sdlFileName, Integer iOffset, Integer iLength) throws SdlException 
+	public void PutFileStream(InputStream is, String sdlFileName, Long iOffset, Long iLength) throws SdlException 
 	{
 		PutFile msg = RPCRequestFactory.buildPutFile(sdlFileName, iOffset, iLength);		
 		startRPCStream(is, msg);
@@ -4299,7 +4297,7 @@ public abstract class SdlProxyBase<proxyListenerType extends IProxyListenerBase>
 	 * @return OutputStream - The output stream of byte data that is written to by the app developer
 	 * @throws SdlException
 	*/
-	public OutputStream PutFileStream(String sdlFileName, Integer iOffset, Integer iLength) throws SdlException 
+	public OutputStream PutFileStream(String sdlFileName, Long iOffset, Long iLength) throws SdlException 
 	{
 		PutFile msg = RPCRequestFactory.buildPutFile(sdlFileName, iOffset, iLength);		
 		return startRPCStream(msg);
@@ -4319,7 +4317,7 @@ public abstract class SdlProxyBase<proxyListenerType extends IProxyListenerBase>
 	 * @param  bSystemFile - Indicates if the file is meant to be passed thru core to elsewhere on the system.
 	 * @throws SyncException
 	*/
-	public void PutFileStream(InputStream is, String syncFileName, Integer iOffset, Integer iLength, FileType fileType, Boolean bPersistentFile, Boolean bSystemFile) throws SdlException
+	public void PutFileStream(InputStream is, String syncFileName, Long iOffset, Long iLength, FileType fileType, Boolean bPersistentFile, Boolean bSystemFile) throws SdlException
 	{
 		PutFile msg = RPCRequestFactory.buildPutFile(syncFileName, iOffset, iLength, fileType, bPersistentFile, bSystemFile);
 		startRPCStream(is, msg);
@@ -4339,7 +4337,7 @@ public abstract class SdlProxyBase<proxyListenerType extends IProxyListenerBase>
 	 * @return OutputStream - The output stream of byte data that is written to by the app developer
 	 * @throws SyncException
 	*/
-	public OutputStream PutFileStream(String sdlFileName, Integer iOffset, Integer iLength, FileType fileType, Boolean bPersistentFile, Boolean bSystemFile) throws SdlException
+	public OutputStream PutFileStream(String sdlFileName, Long iOffset, Long iLength, FileType fileType, Boolean bPersistentFile, Boolean bSystemFile) throws SdlException
 	{
 		PutFile msg = RPCRequestFactory.buildPutFile(sdlFileName, iOffset, iLength, fileType, bPersistentFile, bSystemFile);
 		return startRPCStream(msg);
@@ -4360,9 +4358,9 @@ public abstract class SdlProxyBase<proxyListenerType extends IProxyListenerBase>
 	 * @return boolean - True if the putfile stream was started successfully, false if an exception occurred during stream creation. 
 	 * @throws SdlException
 	*/	
-	public boolean PutFileStream(String sPath, String sdlFileName, Integer iOffset, FileType fileType, Boolean bPersistentFile, Boolean bSystemFile, Integer iCorrelationID) throws SdlException 
+	public boolean PutFileStream(String sPath, String sdlFileName, Long iOffset, FileType fileType, Boolean bPersistentFile, Boolean bSystemFile, Integer iCorrelationID) throws SdlException 
 	{
-		PutFile msg = RPCRequestFactory.buildPutFile(sdlFileName, iOffset, 0, fileType, bPersistentFile, bSystemFile, iCorrelationID);
+		PutFile msg = RPCRequestFactory.buildPutFile(sdlFileName, iOffset, Long.valueOf((long)0), fileType, bPersistentFile, bSystemFile, iCorrelationID);
 		return startPutFileStream(sPath, msg);
 	}
